rename: scripts/audit.py → scripts/lint.py, test_audit.py → test_lint.py
